bot/requirements.txt
aiogram
google-genai
python-dotenv
markdown
aiofiles
fastapi
uvicorn
aiosqlite
gTTS
